pyproject.toml
[project]
name = "rvr-chat"
version = "0.1.0"
description = "Local LLM chat interface for Sphero RVR via MCP"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Jason Scott Person", email = "jsperson+rvr@gmail.com" }
]
keywords = [
    "sphero",
    "rvr",
    "ollama",
    "llm",
    "mcp",
    "robotics",
    "raspberry-pi",
]
dependencies = [
    "ollama>=0.4.0",
    "pyyaml>=6.0",
    "prompt_toolkit>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
rvr-chat = "rvr_chat.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rvr_chat"]
